+++ x/inflation/types/interfaces.go
@@ -42,4 +42,5 @@ type StakingKeeper interface {
 
 type SudoKeeper interface {
 	GetRootAddr(ctx sdk.Context) (sdk.AccAddress, error)
+	CheckPermissions(contract sdk.AccAddress, ctx sdk.Context) error
 }